Add quick setup option to clear map
