+++ src/debug_menu_types.h
@@ -105,6 +105,7 @@ enum class debug_menu_index : int {
     EXPORT_SELF,
     QUICK_SETUP,
     QUICK_SETUP_FLAG_DIRTY,
+    QUICK_SETUP_CLEAR_MAP,
     TOGGLE_SETUP_MUTATION,
     NORMALIZE_BODY_STAT,
     SIX_MILLION_DOLLAR_SURVIVOR,
