+++ tests/char_sight_test.cpp
@@ -12,6 +12,7 @@
 #include "lightmap.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_scale_constants.h"
 #include "memory_fast.h"
 #include "npc.h"
