+++ tests/char_suffer_test.cpp
@@ -16,6 +16,7 @@
 #include "item_location.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "options_helpers.h"
 #include "player_helpers.h"
 #include "point.h"
