+++ tests/crafting_test.cpp
@@ -34,6 +34,7 @@
 #include "itype.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_iterator.h"
 #include "map_selector.h"
 #include "mapdata.h"
