+++ tests/degradation_test.cpp
@@ -16,6 +16,7 @@
 #include "iuse.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_scale_constants.h"
 #include "map_selector.h"
 #include "player_activity.h"
