+++ tests/effect_test.cpp
@@ -17,6 +17,7 @@
 #include "item.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "monster.h"
 #include "npc.h"
 #include "player_helpers.h"
