+++ tests/enchantments_test.cpp
@@ -16,6 +16,7 @@
 #include "item_location.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "messages.h"
 #include "monster.h"
 #include "player_helpers.h"
