+++ tests/firstaid_test.cpp
@@ -13,6 +13,7 @@
 #include "item_location.h"
 #include "itype.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "npc.h"
 #include "player_activity.h"
 #include "player_helpers.h"
