+++ tests/flying_creature_test.cpp
@@ -7,6 +7,7 @@
 #include "game.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "monster.h"
 #include "player_helpers.h"
 #include "point.h"
