+++ tests/limb_test.cpp
@@ -16,6 +16,7 @@
 #include "magic_enchantment.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "npc.h"
 #include "npc_opinion.h"
 #include "options_helpers.h"
