+++ tests/map_path_test.cpp
@@ -11,6 +11,7 @@
 #include "game.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_iterator.h"
 #include "monster.h"
 #include "pathfinding.h"
