+++ tests/map_test.cpp
@@ -22,6 +22,7 @@
 #include "itype.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_scale_constants.h"
 #include "map_selector.h"
 #include "monster.h"
