+++ tests/mapgen_remove_npcs_test.cpp
@@ -11,6 +11,7 @@
 #include "game.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_scale_constants.h"
 #include "mapgen_helpers.h"
 #include "npc.h"
