+++ tests/martial_art_test.cpp
@@ -15,6 +15,7 @@
 #include "item.h"
 #include "magic_enchantment.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_scale_constants.h"
 #include "martialarts.h"
 #include "monster.h"
