+++ tests/monster_stair_test.cpp
@@ -10,6 +10,7 @@
 #include "game.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "monster.h"
 #include "point.h"
 #include "type_id.h"
