+++ tests/monster_test.cpp
@@ -28,6 +28,7 @@
 #include "line.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_iterator.h"
 #include "map_scale_constants.h"
 #include "mapdata.h"
