+++ tests/monster_vision_test.cpp
@@ -5,6 +5,7 @@
 #include "coordinates.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "monster.h"
 #include "point.h"
 #include "type_id.h"
