+++ tests/morale_test.cpp
@@ -15,6 +15,7 @@
 #include "item_location.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "morale.h"
 #include "npc.h"
 #include "pimpl.h"
