+++ tests/overmap_test.cpp
@@ -29,7 +29,7 @@
 #include "item_factory.h"
 #include "itype.h"
 #include "map.h"
-#include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_iterator.h"
 #include "map_scale_constants.h"
 #include "mapbuffer.h"
