+++ tests/player_activities_test.cpp
@@ -26,6 +26,7 @@
 #include "iuse_actor.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "mapdata.h"
 #include "monster.h"
 #include "monster_helpers.h"
