+++ tests/reading_test.cpp
@@ -18,6 +18,7 @@
 #include "itype.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "player_activity.h"
 #include "player_helpers.h"
 #include "pocket_type.h"
