+++ tests/swim_under_test.cpp
@@ -14,6 +14,7 @@
 #include "item.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "mapdata.h"
 #include "monster.h"
 #include "npc.h"
