+++ tests/throwing_test.cpp
@@ -15,6 +15,7 @@
 #include "item.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "monster.h"
 #include "npc.h"
 #include "player_helpers.h"
