+++ tests/vehicle_efficiency_test.cpp
@@ -18,6 +18,7 @@
 #include "itype.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "point.h"
 #include "test_data.h"
 #include "test_statistics.h"
