+++ tests/vehicle_fake_part_test.cpp
@@ -12,6 +12,7 @@
 #include "enums.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "pathfinding.h"
 #include "player_helpers.h"
 #include "point.h"
