+++ tests/vehicle_part_test.cpp
@@ -21,6 +21,7 @@
 #include "itype.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "player_activity.h"
 #include "player_helpers.h"
 #include "point.h"
