+++ tests/vehicle_power_test.cpp
@@ -22,6 +22,7 @@
 #include "level_cache.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_scale_constants.h"
 #include "options_helpers.h"
 #include "player_helpers.h"
