+++ tests/vision_test.cpp
@@ -17,6 +17,7 @@
 #include "game.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_scale_constants.h"
 #include "map_test_case.h"
 #include "monster.h"
