+++ tests/visitable_zone_test.cpp
@@ -10,6 +10,7 @@
 #include "cuboid_rectangle.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "map_iterator.h"
 #include "map_scale_constants.h"
 #include "monster.h"
