+++ tests/widget_test.cpp
@@ -27,6 +27,7 @@
 #include "magic.h"
 #include "map.h"
 #include "map_helpers.h"
+#include "map_helpers_tests.h"
 #include "mission.h"
 #include "monster.h"
 #include "options_helpers.h"
